Unreviewed, skip misc-bugs-847389-jpeg2000.js in jsc stress tests https://bugs.webkit.org/show_bug.cgi?id=159334
In debug build, it takes too much time.
Since run-jsc-stress-tests can take the arbitrary JSC shell, ifJSCArgIsntProvidedAreWeReleaseBuild
is not good source to determine whether the test is executed with the debug built JSC shell.
So here, we skip this test.
This test is still executed under the run-webkit-tests. So the coverage is still kept.

* js/regress/script-tests/misc-bugs-847389-jpeg2000.js:

Identifier: 177536@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202805 268f45cc-cd09-0410-ab3c-d52691b4dbfc