Use Radical* constants from the OpenType MATH table. https://bugs.webkit.org/show_bug.cgi?id=155638
Patch by Frederic Wang <fwang@igalia.com> on 2016-07-08
Reviewed by Brent Fulgham.

Source/WebCore:

Test: mathml/mathml-in-html5/root-parameters-1.html

We make the radical vertical gap depends on displaystyle.
This is the only remaining step to use all the Radical* constants from the MATH table.
We also introduce a ruleThicknessFallback function for future use.

* rendering/mathml/RenderMathMLBlock.h:
(WebCore::RenderMathMLBlock::ruleThicknessFallback): Add this helper function since that
calculation is used in several places.
* rendering/mathml/RenderMathMLRoot.cpp:
(WebCore::RenderMathMLRoot::updateStyle): Reorganize the way we set constant parameters,
add more comments and take into account the displaystyle for the vertical gap.

LayoutTests:

Import a test from the MathML in HTML5 test suite to verify all the Radical* parameters.
For test RadicalDegreeBottomRaisePercent, WebKit's drawing of one of the radical sign
is weird. This causes unexpected height of the radical and a fortiori wrong vertical
position of the index. For now, we tweak that height to make the test pass.

* imported/mathml-in-html5/fonts/radical-degreebottomraisepercent25-rulethickness1000.woff: Added.
* imported/mathml-in-html5/fonts/radical-displaystyleverticalgap7000-rulethickness1000.woff: Added.
* imported/mathml-in-html5/fonts/radical-extraascender3000-rulethickness1000.woff: Added.
* imported/mathml-in-html5/fonts/radical-kernafterdegreeminus5000-rulethickness1000.woff: Added.
* imported/mathml-in-html5/fonts/radical-kernbeforedegree4000-rulethickness1000.woff: Added.
* imported/mathml-in-html5/fonts/radical-rulethickness8000.woff: Added.
* imported/mathml-in-html5/fonts/radical-verticalgap6000-rulethickness1000.woff: Added.
* imported/mathml-in-html5/mathml/presentation-markup/radicals/root-parameters-1-expected.txt: Added.
* imported/mathml-in-html5/mathml/presentation-markup/radicals/root-parameters-1.html: Added.
* platform/gtk/mathml/radical-fallback-expected.png: Updated to take into account new rendering.
* platform/gtk/mathml/radical-fallback-expected.txt: Ditto.
* platform/mac/mathml/presentation/mo-stretch-expected.txt: Ditto.
* platform/mac/mathml/presentation/roots-expected.txt: Ditto.
* platform/mac/mathml/radical-fallback-expected.txt: Ditto.
* platform/ios-simulator/mathml/presentation/mo-stretch-expected.txt: Ditto.
* platform/ios-simulator/mathml/presentation/roots-expected.txt: Ditto.
* platform/ios-simulator/mathml/radical-fallback-expected.txt: Ditto.

Identifier: 177698@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202977 268f45cc-cd09-0410-ab3c-d52691b4dbfc