v2: WebContent crash due to RELEASE_ASSERT(!m_inLoadPendingImages) in StyleResolver::~StyleResolver() https://bugs.webkit.org/show_bug.cgi?id=159722
Reviewed by Andreas Kling.

We have crashes where a StyleResolver is deleted underneath pseudoStyleForElement (key parts of the stack):

0   WebCore::StyleResolver::~StyleResolver
3   WebCore::AuthorStyleSheets::updateActiveStyleSheets
4   WebCore::Document::styleResolverChanged
5   WebKit::WebPage::viewportConfigurationChanged()
6   WebKit::WebPage::mainFrameDidLayout()
9   WebCore::FrameLoader::checkCompleted
13  WebCore::ResourceLoader::cancel
19  WebKit::WebLoaderStrategy::loadResource
24  WebCore::Style::loadPendingImage
27  WebCore::StyleResolver::pseudoStyleForElement
29  WebCore::RenderTreeUpdater::updateBeforeOrAfterPseudoElement
33  WebCore::Document::recalcStyle

This appears to be happening when a content blocker blocks a resource load for an image referenced from a stylesheet
and triggers synchronous cancellation of the load. With engine in suitable state this can clear style resolver.

No test, don't know how to make one. This is very timing and engine state dependent.

* dom/AuthorStyleSheets.cpp:
(WebCore::AuthorStyleSheets::updateActiveStyleSheets):

We have an existing check here that prevents destruction of the style resolver when we are in the middle of
a style resolution. However the old inStyleRecalc() bit no longer covers the render tree update phase. Pseudo
elements are resolved during render tree update.

Fix by adding a check for inRenderTreeUpdate() bit too.

This just fixes a regression. A proper fix would be to gather all resources during style resolution
and trigger the loads afterwards.



Identifier: 177872@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203172 268f45cc-cd09-0410-ab3c-d52691b4dbfc