Add DejaVu Math TeX Gyre to the list of math fonts. https://bugs.webkit.org/show_bug.cgi?id=159805
Patch by Frederic Wang <fwang@igalia.com> on 2016-07-15
Reviewed by Brent Fulgham.

DejaVu 2.36 has a new math font that can be used for MathML rendering. Because this font is
likely to be installed on many systems (Linux, LibreOffice, etc) we include it in the default
list of font-families in mathml.css in order to increase the chance to find a math font.

No new tests, it only affects rendering when DejaVu Math TeX Gyre is installed on the system.

* css/mathml.css:
(math):

Identifier: 177969@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203279 268f45cc-cd09-0410-ab3c-d52691b4dbfc