[mac] LayoutTest fast/css/ancestor-of-hovered-element-detached.html is flaky https://bugs.webkit.org/show_bug.cgi?id=159173
Reviewed by Alexey Proskuryakov.

The test ancestor-of-hovered-element-detached.html is not reliably
covering r202324.

In the case of that test, the first style update is done with
the flag TeardownType::KeepHoverAndActive. This keeps the ":hover" state
set despite the subtree losing their renderer.
This looks like a serious issue that should be tested separately.

* fast/css/ancestor-of-hovered-element-removed.html:
The :hover update timer is actually a zero timer.
It is reasonable to expect a full update after one frame update.

* fast/css/ancestor-of-hovered-element-detached-expected.txt: Removed.
* fast/css/ancestor-of-hovered-element-detached.html: Removed.
* platform/ios-simulator/TestExpectations:
* platform/mac/TestExpectations:


Identifier: 178011@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203321 268f45cc-cd09-0410-ab3c-d52691b4dbfc