Add regression test for r203392 https://bugs.webkit.org/show_bug.cgi?id=159949
Reviewed by Brady Eidson.

* TestWebKitAPI/Tests/WebKit2Cocoa/LocalStorageClear.mm:
(fileSize):
(swizzledBundleIdentifierWebBookmarksD):
(defaultApplicationCacheDirectory):
(TEST):
(swizzledBundleIdentifierMobileSafari): Deleted.
This actually tests that webbookmarksd uses the path quirk in WebsiteDataStore::defaultApplicationCacheDirectory.
I wanted to verify that it uses the same path as MobileSafari, but swizzling out the bundleIdentifier twice caused
problems with the static bools that are set once in RuntimeApplicationChecks.mm.



Identifier: 178112@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203429 268f45cc-cd09-0410-ab3c-d52691b4dbfc